% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kospacing.R
\name{spacing}
\alias{spacing}
\title{do korean spacing}
\usage{
spacing(ko_sents)
}
\arguments{
\item{ko_sents}{korean sentences (limit of sentence length  is 198, must be splitted.)}
}
\value{
auto-spaced korean sentences
}
\description{
do korean spacing
}
